Наполовину реализована система составления плана чтения (Резервное сохранение).

+++ includes/lib.php
@@ -1,5 +1,5 @@
 <?php
-require 'db.php';
+require 'reader-file.php';
 if (isset($_COOKIE['logged_user']) && isset($_COOKIE['key']))
     if (checkKey($_COOKIE['key']))
 switch ($_POST['function']){
@@ -52,7 +52,8 @@ switch ($_POST['function']){
                 $last[3]=0;
                 B::updateBase('users_info', array('files_disk_space', 'last_books'), array($size, json_encode($last)), array('login'), array($_COOKIE['logged_user']));
                 removeDirectory($path);
-                unlink('../'.$data[0]['cover']);
+                if (stripos($data[0]['cover'],"img/epub.png")===false && stripos($data[0]['cover'],'img/fb2.png')===false && stripos($data[0]['cover'],'img/pdf.png')===false && stripos($data[0]['cover'],'img/txt.png')===false && stripos($data[0]['cover'],'img/html.png')===false)
+                    unlink('../'.$data[0]['cover']);
                 B::deleteFromBase('users_files', array('id'), array($data[0]['id']));
                 B::deleteFromBase('bookmarks', array('id_book'), array($data[0]['id']));
             }
@@ -88,7 +89,8 @@ switch ($_POST['function']){
                 $last[3]=0;
                 B::updateBase('users_info', array('files_disk_space', 'last_books'), array($size, json_encode($last)), array('login'), array($_COOKIE['logged_user']));
                 removeDirectory($path);
-                unlink('../'.$data[0]['cover']);
+                if (stripos($data[0]['cover'],"img/epub.png")===false && stripos($data[0]['cover'],'img/fb2.png')===false && stripos($data[0]['cover'],'img/pdf.png')===false && stripos($data[0]['cover'],'img/txt.png')===false && stripos($data[0]['cover'],'img/html.png')===false)
+                    unlink('../'.$data[0]['cover']);
                 B::deleteFromBase('users_files', array('id'), array($data[0]['id']));
                 B::deleteFromBase('bookmarks', array('id_book'), array($data[0]['id']));
                 echo getBodyMain();
@@ -96,7 +98,6 @@ switch ($_POST['function']){
         } else {
             delCookies('logged_user');
             delCookies('key');
-            echo '<META HTTP-EQUIV="Refresh" CONTENT="0; URL=/">';
         }
         break;
     case 'markAllNotificationsAsRead':
@@ -139,11 +140,169 @@ switch ($_POST['function']){
   <div class="panel-body">
     '.$item['notification'].'
   </div>
-  <div class="panel-footer"><div class="btn-group btn-group-sm"><button class="btn btn-success btn-bookmark btn-sm btn-rad" id="btnReadNotification'.$item['id'].'" onclick="readNotification('.$item['id'].')">Read</button></div></div>
+  <div class="panel-footer"><div class="btn-group btn-group-sm"><button class="btn btn-success btn-bookmark btn-sm btn-rad" id="btnReadNotification'.$item['id'].'" disabled="disabled" onclick="readNotification('.$item['id'].')">Read</button></div></div>
 </div>';
         }
         echo $modalbody;
         break;
+    case 'checkForm':
+        $stmt=B::selectFromBase('reading_plans',null,array('id_book'),array($_POST['id']));
+        $data=$stmt->fetchAll();
+        $response='';
+        if (count($data)==0){
+            $response=json_encode(array('text' => 'Select the date to which you want to finish reading '.$_POST['name'].' and the system will make you a reading plan:', 'value' => '0'));
+        } else {
+            $json = json_decode($data[0]['progress'], true);
+            $datetime1 = date_create($json['today_date']);
+            $datetime2 = date_create(date('Y-m-d'));
+            $interval = date_diff($datetime1, $datetime2);
+            if ($interval->days > 0) {
+                $new_json = json_encode(array("must_read" => $json['must_read'], "today_read" => $json['today_read'], "today_date" => date('Y-m-d')));
+                $readingPlan = new readingPlan();
+                $stmt = B::selectFromBase('users_files', null, array('id'), array($data[0]['id_book']));
+                $dbook = $stmt->fetchAll();
+                $json = json_decode($dbook[0]['progress'], true);
+                $file_info = pathinfo($dbook[0]['path']);
+                $infoRes='';
+                switch ($file_info['extension']){
+                    case 'epub':
+                        $readingPlan->getEPUBFB2Plan($json, $data[0]['date']);
+                        $infoRes = $readingPlan->infoRes;
+                        break;
+                    case 'fb2':
+                        $readingPlan->getEPUBFB2Plan($json, $data[0]['date']);
+                        $infoRes = $readingPlan->infoRes;
+                        break;
+                    case 'pdf':
+
+                        break;
+                    case 'txt':
+
+                        break;
+                    case 'html':
+
+                        break;
+                }
+                B::updateBase('reading_plans', array('progress', 'info_res'), array($new_json, $infoRes), array('id_book'), array($_POST['id']));
+                $datetime1 = date_create($data[0]['date']);
+                $datetime2 = date_create(date('Y-n-j'));
+                $interval = date_diff($datetime1, $datetime2);
+                $response = json_encode(array('value' => '1', 'interval' => $data[0]['number_days'], 'daysLeft' => (string)$interval->days, 'progress' => $json['progress'], 'infoRes' => $infoRes));
+            } else {
+                $stmt = B::selectFromBase('users_files', null, array('id'), array($data[0]['id_book']));
+                $dbook = $stmt->fetchAll();
+                $json = json_decode($dbook[0]['progress'], true);
+                $datetime1 = date_create($data[0]['date']);
+                $datetime2 = date_create(date('Y-n-j'));
+                $interval = date_diff($datetime1, $datetime2);
+                $response = json_encode(array('value' => '1', 'interval' => $data[0]['number_days'], 'daysLeft' => (string)$interval->days, 'progress' => $json['progress'], 'infoRes' => $data[0]['info_res']));
+            }
+        }
+        echo $response;
+        break;
+    case 'setTime':
+        $date=explode('.',$_POST['date']);
+        $response='';
+        $infoTerm='';
+        if (checkdate($date[1],$date[0],$date[2])) {
+            $stmt=B::selectFromBase('users',null,array('login'),array($_COOKIE['logged_user']));
+            $data=$stmt->fetchAll();
+            $id_user=$data[0]['id'];
+            $stmt=B::selectFromBase('users_files',null,array('id_user','id'),array($id_user,$_POST['id']));
+            $data=$stmt->fetchAll();
+            if (count($data)>0){
+                $dates="$date[2]-$date[1]-$date[0]";
+                $file_info = pathinfo($data[0]['path']);
+                $day_norm='';
+                $infoTerm='Your plan is created.';
+                $infoRes='You need to read';
+                $interval=0;
+                $json=json_decode($data[0]['progress'],true);
+                switch ($file_info['extension']){
+                    case 'epub':
+                        $readingPlan= new readingPlan();
+                        $readingPlan->getEPUBFB2Plan($json,$dates);
+                        $infoRes=$readingPlan->infoRes;
+                        $day_norm=$readingPlan->day_norm;
+                        $interval=$readingPlan->interval;
+                        break;
+                    case  'fb2':
+                        $readingPlan= new readingPlan();
+                        $readingPlan->getEPUBFB2Plan($json,$dates);
+                        $infoRes=$readingPlan->infoRes;
+                        $day_norm=$readingPlan->day_norm;
+                        $interval=$readingPlan->interval;
+                        break;
+                    case  'pdf':
+
+                        break;
+                    case 'txt':
+
+                        break;
+                    case  'html':
+
+                        break;
+                }
+                $progress=json_encode(array("must_read" => $day_norm, "today_read" => "0", "today_date" => date('Y-m-d')));
+                B::inBase('reading_plans',array('id_book','id_user','progress','date','number_days','info_res'),array($_POST['id'],$id_user,$progress,$dates,$interval,$infoRes));
+                $response=json_encode(array("correct" => "1", "otherInfo" => $infoTerm, "infoRes" => $infoRes, "progress" => $json['progress'],"numberDays" => $interval));
+            }
+        } else {
+            $response=json_encode(array("correct" => "0", "otherInfo" => $infoTerm));
+        }
+        echo $response;
+        break;
+}
+class readingPlan{
+    var $infoRes;
+    var $day_norm;
+    var $interval;
+    function getEPUBFB2Plan($json,$dates){
+        $start = date_create(date('Y-m-d'));
+        $end = date_create($dates);
+        $day_s='';
+        $interval = date_diff($start, $end);
+        if ($interval->days==0) {
+            $interval = 1;
+            $day_s = 'day';
+        }
+        else {
+            $interval = $interval->days;
+            $day_s='days';
+        }
+        $infoRes='You need to read';
+        $day_norm='';
+        $resources=new otherRousorces();
+        $res=$resources->getOtherInfo();
+        $numberOfMinutes=round(($json['p']-$json['p']/100*$json['progress'])/$interval/$res['speedReading']/$res['lengthOfWord'],0,PHP_ROUND_HALF_UP);
+        if ($numberOfMinutes<2){
+            $infoRes.=" less than a minute a day, and in $interval $day_s you will finish reading this book.";
+        } else {
+            if ($numberOfMinutes<60) {
+                $infoRes .= " about $numberOfMinutes minutes a day, and in $interval $day_s you will finish reading this book.";
+            }
+            else {
+                if (round($numberOfMinutes/60,0,PHP_ROUND_HALF_UP)>=2)
+                    $hour_s='hours';
+                else
+                    $hour_s='hour';
+                $hours=round($numberOfMinutes/60,0,PHP_ROUND_HALF_UP);
+                if (abs((int)($numberOfMinutes/60)-round($numberOfMinutes/60,1,PHP_ROUND_HALF_UP))<0.5) {
+                    $infoRes .=" about $hours $hour_s a day, and in $interval $day_s you will finish reading this book.";
+                }
+                else {
+                    $infoRes .=" about $hours and half $hour_s a day, and in $interval $day_s you will finish reading this book.";
+                }
+            }
+        }
+        if ($numberOfMinutes==0)
+            $day_norm=$res['speedReading'];
+        else
+            $day_norm=$res['speedReading']*$numberOfMinutes;
+        $this->infoRes=$infoRes;
+        $this->day_norm=$day_norm;
+        $this->interval=$interval;
+    }
 }
 class showBook{
     function showAllBook(){
